Move x509proxy module alongside rest module
